# Removed `OfType` usage. Less allocations and much faster