@En3Tho 2021-02-07
I don't think it's necessary to replace easily readable Linq with manual foreach in methods that run only once. I doubt it will bring any noticeable difference