@MCCshreyas 2021-02-07
LINQ brings readability, but it lowers the performance! And this method is get's called at startup, and we want to make startup as fast as possible! 
Some more info -> https://github.com/dotnet/runtime/discussions/45060