@AmrAlSayed0 2021-02-07
You're test is wrong. You're list has objects of type ```Type``` and you're trying to cast it to ```string``` in the first test.

The second test just checks for equality which is not the same thing.