@javiercn 2021-02-10
I agree with @stephentoub that the code is fine as it is and we don't need to change it unless we see significant performance benefits. I believe there should be other areas we should look at first before going for these micro optimizations.