@MCCshreyas 2021-02-10
My intension is to remove out slow things that impacts startup. I know resulting code is not quite readable, but also it's not that much hard to understand! 😄