@javiercn 2021-02-10
> My intension is to remove out slow things that impacts startup.

What's the improvement over the overall startup time?