@sharwell 2021-02-10
@javiercn would you agree that the alternative form I proposed is equally readable to the original?