@javiercn 2021-02-10
> @javiercn would you agree that the alternative form I proposed is equally readable to the original?

Yes, I'm not really complaining about any of these being less readable, I'm suggesting that maybe is not worth changing given that I don't have a clear idea of how much improvement this brings to the overall startup experience (I see from the issue linked that there is a general effort to improve this area in general).

I'm just pointing out that I don't know the impact this has in total startup time, but I suspect is negligible, and at that point, is it worth changing the code?

The risk here is the chance of introducing a small/subtle behavior change/bug, if we are taking that risk, at least we should make sure is worth it.

Again, I'm happy if we take this PR in whatever shape provided the implementation remains equivalent, but for future reference we should at least establish a minimal threshold to determine what's worth optimizing.

Does that make sense?