@sharwell 2021-02-10
> Does that make sense?

Certainly, was just getting clarification on the reason for wanting more data. 👍 

@MCCshreyas if you want to capture more data about this, one of the easiest things to capture is a count of the number of times this method is called during the startup process. If the count varies according to the program configuration, then a set of minimum/expected/maximum counts would be particularly helpful.