@pranavkm 2021-02-10
> f you want to capture more data about this, one of the easiest things to capture is a count of the number of times this method is called during the startup process

In the templates, it's usually invoked once.