@pranavkm 2021-02-19
@MCCshreyas thanks for your PR. I think we need to be able to notice measurable startup improvements for changes like this. Unfortunately microbenchmarks might not necessarily demonstrate this. I'm closing this. If you get the opportunity to measure startup perf (https://github.com/dotnet/crank has features that can let you do this), we'd be happy to reconsider this change.